Catch two exceptions, some code tweaks

+++ branchrename.py
@@ -6,8 +6,8 @@ import settings
 g = Github(settings.GITHUB_TOKEN)
 
 
-oldrepo = 'master'
-newrepo = settings.BRANCH
+existing_branch_name = 'master'
+new_branch_name = settings.BRANCH
 
 ulogin = g.get_user().login
 
@@ -16,30 +16,37 @@ for repo in g.get_user().get_repos():
   if ulogin == repo.owner.login and repo.fork == False:
 
     try:
-      branch = repo.get_branch(branch=oldrepo)
+      branch = repo.get_branch(branch=existing_branch_name)
 
     except GithubException:
       continue # We expect this if there's no master branch.
 
-    print(repo.name)
+    print(repo.owner.login + "/" + repo.name, end=" ")
+
+    if repo.archived:
+      print('Repository is archived so branch name cannot be changed.')
+      continue
 
     # If we already have a branch named after the new branch, skip this repo.
     try:
-      branch = repo.get_branch(branch=newrepo)
+      branch = repo.get_branch(branch=new_branch_name)
       if branch:
-        print('Branch %s already exists, skipping' % newrepo)
+        print('Branch "%s" already exists, skipping' % new_branch_name)
         continue
 
     except GithubException:
       pass # We expect this.
 
-    src = repo.get_git_ref('heads/%s' % oldrepo)
-    print(src.object.sha)
+    if branch.protected:
+      print('Branch "%s" is protected, skipping' % existing_branch_name)
+      continue
 
-    repo.create_git_ref('refs/heads/%s' % newrepo, sha=src.object.sha)
+    src = repo.get_git_ref('heads/%s' % existing_branch_name)
 
+    print(src.object.sha, existing_branch_name, "=>", new_branch_name)
+
+    # Create a new brach that points to the same commit as the existing branch.
+    repo.create_git_ref('refs/heads/%s' % new_branch_name, sha=src.object.sha)
+
+    # Delete the old branch.
     src.delete()
-
-    print('Done')
-    print()
-
